+++ authentication/authentication-core/src/main/java/io/mosip/authentication/core/partner/dto/PartnerCurrentBalanceDTO.java
@@ -1,7 +1,5 @@
 package io.mosip.authentication.core.partner.dto;
 
-import java.math.BigDecimal;
-
 import lombok.Data;
 
 /**
@@ -14,5 +12,5 @@ public class PartnerCurrentBalanceDTO {
 
 	private String partnerId;
 
-	private BigDecimal balance;
+	private Double balance;
 }
